+++ changelog.md
@@ -13,6 +13,10 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
   should now be changed to `require('task-scheduler-collection/single-instance')`.
 - Change constructor signiture of `SingleInstanceTaskScheduler` to make it easier to use.
 
+### Removed
+- Marked function `buildEvaluator` in `single-instance` module as private by adding
+  underscore prefix. Export will be removed in future.
+
 ## [0.2.0] - 2021-05-25
 ### Added
 - Added `SingleInstanceTaskScheduler.nextRunTime` property.